Delete dead code from tetwild, triwild and simwild
2687 lines removed, nothing added. Every function deleted here has zero callers,
or its only callers were themselves dead; every member deleted is either never
read or never written. This is subtraction before a de-duplication pass -- there
is no point factoring code that nothing runs.

  586  tetwild insertion_by_volumeremesher_old
  490  adjust_sizing_field_serial, in all four meshes
  381  simwild embed_surface_old
  373  tetwild's dead debug writers and predicate copies: output_embedded_polygon_mesh,
       output_embedded_polygon_surface_mesh, output_tetrahedralized_embedded_mesh,
       output_init_tetmesh, output_tracked_surface, check_polygon_face_validity,
       check_nondegenerate_tets, orient3D, orient3D_wmtk_rational,
       checkTrackedFaces_wmtk_rational
  358  TriangleInsertion.cpp, 469 lines down to 111
  115  triangulate_polygon_face (94% duplicated tetwild/simwild) + two dangling
       declarations, one of which had no definition at all
  ~120 dead members: v1_param_type/v2_param_type and the is_limit_length field in
       all four operation caches, tetwild's nine never-written CollapseInfoCache
       debug fields, check_vertex_param_type, the write-only Parameters::output_path
       in tetwild and triwild and Parameters::max_its in simwild, and the
       v1_id/v2_id/sizing_ratio prelude computed and discarded in all four collapse
       is_weight_up_to_date lambdas
  ~260 stale comment blocks and commented-out legacy code

TriangleInsertion.cpp lost 76% of its content -- init_from_input_surface,
finalize_triangle_insertion, internal_insert_single_triangle and the
triangle_insertion_before/after overrides, none reachable since the arrangement
replaced that path. Only init_from_delaunay_box_mesh survives, so the file is
renamed DelaunayBoxMesh.cpp to say what it now contains.

Deleting insertion_by_volumeremesher_old left the surviving
insertion_by_volumeremesher documented entirely as a diff against it -- five
"KEY DIFFERENCE vs _old" blocks and a banner that opened by naming it. Rewritten
to describe what the function does. Same for two sizing-field clamp comments that
cited adjust_sizing_field_serial as their justification.

Validation, per the policy that tetwild and triwild must not move:

  * ctest 107/107.
  * All 53 registered integration configs byte-identical on #v, #t, max_energy,
    avg_energy, all_rounded, hausdorff, coverage and #iterations.
  * All 30 hidden [challenging] models byte-identical on the same fields.

Two caveats worth recording for the next pass. tetwild_crown and tetwild_octocat
run at num_threads=10 and DID differ -- but they differ from themselves: the same
binary gave crown 5734 then 5673 vertices and octocat 7746 then 7854 on
consecutive runs. The parallel scheduler is non-deterministic, as
tests/python/test_integration.py documents. Both are byte-identical when re-run
at num_threads=0, which is the only setting where comparing them means anything.

And find_open_boundary looks dead if the reachability grep excludes its own file;
it is called from init_from_Volumeremesher and is live. Grep including internal
references before deleting anything else here.
